🚚 [RUMF-1379] rename module getSelectorsFromElement
Since we won't need multiple selectors anymore, rename the module to
singular 'getSelectorFromElement'
