+++ packages/rum-core/src/domain/rumEventsCollection/action/getSelectorFromElement.ts
