+++ packages/rum-core/src/domain/rumEventsCollection/action/trackClickActions.ts
@@ -22,7 +22,7 @@ import { waitPageActivityEnd } from '../../waitPageActivityEnd'
 import type { ClickChain } from './clickChain'
 import { createClickChain } from './clickChain'
 import { getActionNameFromElement } from './getActionNameFromElement'
-import { getSelectorsFromElement } from './getSelectorsFromElement'
+import { getSelectorsFromElement } from './getSelectorFromElement'
 import type { MouseEventOnElement, GetUserActivity } from './listenActionEvents'
 import { listenActionEvents } from './listenActionEvents'
 import { computeFrustration } from './computeFrustration'
