+++ packages/rum-core/src/index.ts
@@ -35,4 +35,4 @@ export { getMutationObserverConstructor } from './browser/domMutationObservable'
 export { initViewportObservable, getViewportDimension } from './browser/viewportObservable'
 export { RumInitConfiguration, RumConfiguration } from './domain/configuration'
 export { DEFAULT_PROGRAMMATIC_ACTION_NAME_ATTRIBUTE } from './domain/rumEventsCollection/action/getActionNameFromElement'
-export { STABLE_ATTRIBUTES } from './domain/rumEventsCollection/action/getSelectorsFromElement'
+export { STABLE_ATTRIBUTES } from './domain/rumEventsCollection/action/getSelectorFromElement'
